requirements.txt
fastapi
uvicorn
pydantic
pydantic[email]
passlib[bcrypt]
pyjwt
python-multipart
